gemini-lamp-agent/requirements.txt
psycopg2-binary>=2.9.7
python-dotenv>=1.0.0
requests>=2.31.0
langchain>=0.1.0
langchain-google-genai>=1.0.0
